+++ cardano-crypto-leios/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog for `cardano-crypto-leios`
 
-## 0.1.0.2
+## 0.2.0.0
 
-*
+* Remove `encodeLeiosCert`, `decodeLeiosCert`, `encodeBitField` and `decodeBitField`
+* Export `BitField` constructor
 
 ## 0.1.0.1